Refactor: 결과 모달에 ModalLayout 적용 및 하드코딩된 결과 목록 분리
